Full auth test. Test refactor
