+++ tests/shared/__init__.py
@@ -1,0 +1,3 @@
+from .wrappers import *
+
+__all__ = ["Process", "Service", "cleanup"]
